Fix attribute construction in unitaries.U
The construction of unitaries.U relied on overriding the method
_create_array_property() to initialize the correct data structure
in self._matrix. cdef functions can be overridden, but this only
takes effect *after* construction of the object is completed.
Thus, the parent constructor of U, i.e. MatrixOperator.__cinit__(),
calls the method of the parent, not the overridden one. This bug
is fixed by moving the appropriate function calls to the __init__()
method, which is only called after object construction is complete.
This also makes it necessary to manually call the __init__() method
of the parent class in unitaries.U.__init__().