Temporarily fix bug when using random numbers
The pseudorandom seed for the Mersenne Twister used by QuEST and
set in the `core` module was not communicated to the other modules,
so measurements generated identical results in every run, even when
this was not intended. This commit provides a quick temporary fix,
until a more permanent solution is found.