Add borrowing functionality to `Register`
Different objects of type `Register` can now share the same
underlying quest `Qureg`. The borrower is notified when any of the
amplitudes are about to change so it can make a copy of the `Qureg`
if necessary. Borrower and borrowee point at each other with weak
references so they don't keep objects alive longer than necessary.
If a `Register` is about to be destroyed but its `Qureg` has been
borrowed by another `Register`, the `Qureg` is handed off to the
new object instead of being destroyed. This prevents unnecessary
copying of data.