Update ignored folders
Because we now pass `cmake` to `setup_requires`, upon building, some
files are cached by setuptools in the `.eggs` directory at the top
level, which this commit adds to `.gitignore`. It also moves the
folder for non-committable additional files from `extras` to `misc`,
as `extras` might at some point be used for optional dependencies,
following the naming convention of setuptools.