Fix `QuESTError` after switch to dynamic libs
After switching QuEST from a static to a dynamic library, the
previously used method of overriding `invalidQuESTInputError()` is
not working anymore, because the symbol is resolved within the
dynamic library first, thus not calling the overriding symbol in
the Python modules. This commit splits the error handler into two
separate parts. One must be linked into QuEST and overrides the
`invalidQuESTInputError()` method with one that throws a custom
`quest_exception`. The other modifies Cython's exception handling
to be able to handle this custom exception and raise the pyQuEST
exception `QuESTError`.